% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent.r
\name{equiv}
\alias{equiv}
\title{Test to see if two vectors are equivalent.}
\usage{
equiv(x, y, factor_equiv_character = TRUE, ...)
}
\arguments{
\item{x}{the first vector to test for equivalence.}

\item{y}{the second vector to test for equivalence.}

\item{factor_equiv_character}{should factors be treated as equivalend
to characters? (default TRUE)}

\item{...}{other arguments for type-specific equivalence.}
}
\value{
A boolean is returned indicating whether or not the two
vectors are equivalent.
}
\description{
Two numeric vectors are equivalent if their correlation 
coefficient is 1 or -1. Two character or factor vectors are equivalent if 
their is a mapping between labels that makes them the same.
}
\examples{
a <- rnorm(10)
b <- 2 * rnorm(10) + 4
# TRUE because they are the same (up to an affine transformation).
equiv(a, b)

a <- c("a", "a", "b", "c")
b <- c("b", "b", "a", "c")
# TRUE because they are the same (up to a label change).
equiv(a, b)
}
