% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent.r
\name{equiv_columns}
\alias{equiv_columns}
\title{Find equivalent columns in a matrix or data.frame}
\usage{
equiv_columns(x)
}
\arguments{
\item{x}{a matrix or data.frame}
}
\value{
a symmetric boolean matrix where the rows and columns correspond 
to the columns of x and the elements correspond to whether or not the 
columns are equivalent
}
\description{
Test all column combinations to find out which ones are 
equivalent. An upper-triangular matrix is returned with TRUE indicating
columns that are equivalent. Note that the main diagonal along with the
lower triangular values are always FALSE.
}
\examples{

iris$Sepal.Length2 <- 3 * iris$Sepal.Length + 3
equiv_columns(iris)

}
