% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalent.r
\name{has_equiv_column}
\alias{has_equiv_column}
\title{Find the equivalent repeated columns in a matrix or data.frame}
\usage{
has_equiv_column(x, keep_cols = character(), verbose = FALSE)
}
\arguments{
\item{x}{a matrix or data.frame.}

\item{keep_cols}{a character vector of columsn that should be kept.}

\item{verbose}{should information about dropped columns be printed?}
}
\description{
Test the columns from right to left to find out which
ones are duplicated. A vector is returned with TRUE indicating 
columns that are duplicated.
}
